+++ lib/spear/event.ex
@@ -6,7 +6,6 @@ defmodule Spear.Event do
   AppendReq and ReadResp records
   """
 
-  alias Spear.StreamPosition
   require Spear.Records.Streams, as: Streams
   require Spear.Records.Persistent, as: Persistent
   require Spear.Records.Shared, as: Shared
@@ -332,14 +331,6 @@ defmodule Spear.Event do
     |> record_to_map()
     |> from_recorded_event(remaining_opts)
   end
-
-  # coveralls-ignore-start
-  def from_read_response({_type, {position, _position_info}} = read_response, _opts)
-      when position in [:stream_position, :all_stream_position] do
-    StreamPosition.from_read_response(read_response)
-  end
-
-  # coveralls-ignore-stop
 
   @doc """
   Converts an event into a checkpoint